plotCirclize

Using `circlize` package, make a circular visualization of the relationship between genes and corresponding SNVs in genes.

Usage

plotCirclize(
  stana,
  candSp,
  genomeId,
  include_gene = NULL,
  thresh_snp_gene = 5,
  featList = list(),
  cols = c("tomato", "steelblue", "gold", "seagreen"),
  controlColor = "steelblue",
  MAF = FALSE,
  cl = NULL,
  featCircos = list(),
  showGeneName = TRUE,
  returnRawDf = FALSE,
  featThresh = 0,
  cex = 0.3,
  textCex = 0.5,
  bar_width = 10,
  contPalette = c("steelblue", "tomato"),
  discPalette = "Dark2"
)

Arguments

stana

stana object

candSp

candidate species ID

genomeId

genome ID to plot

thresh_snp_gene

include only genes with at least this number of SNVs

featList

features to plot

cols

color of each feature

controlColor

control (not reaching `featThresh`) color in points and barplot

MAF

plot MAF for each group

cl

cluster information (named list)

featCircos

which type of circos to use in the corresponding feature

showGeneName

show gene ID on the plot

returnRawDf

return the raw data frame used in the plot

featThresh

for numeric features, the values above this value will be colored , default to zero.

cex

point size

textCex

text size

bar_width

on circos.barplot

contPalette

palette for continuous scale

discPalette

palette for discrete scale

Value

draw circlize plot